Introduce edisp/tests/test_kernel.py
